+++ src/types/mod.rs
@@ -1,12 +1,14 @@
 use crate::{Env, Value, Result};
 
-pub use {user_ptr::Transfer, vector::Vector};
+pub use {user_ptr::Transfer, vector::Vector, global::GlobalRef};
 
 mod integer;
 mod float;
 mod string;
+
 mod user_ptr;
 mod vector;
+mod global;
 
 // XXX: More accurate would be `CloneFromLisp` or `Decode`, but ...
 /// Converting Lisp [`Value`] into a Rust type.