+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- Added `GlobalRef`, which allows keeping Lisp values around without an `Env`.
 
 ## [0.12.3] - 2020-02-18
 - Added `Value::car`, `Value::cdr`.